gdbus: Allow DBusConnection and DBusMethodInvocation parameters
This is a way for the server implementation to get hold of the
connection that the current method is being invoked on (which may be
important if it's exported on multiple connections, or just to avoid
global variables), for instance to export other objects, or make proxy
calls, on this connection. A full DBusMethodInvocation object can also
be requested in the same way.

Suggested-by: Ivan Molodetskikh <yalterz@gmail.com>